Fixed some comments from the review
 I removed the mandatory flag on POI and station, but I believe it is relevant on addresses.
I solved it by

adding a charset attribute in the elements of the names list
making that list mandatory
removing the 3 special name fields

I propose to take the values from the IRS list, with a default set to "PERSON", so we can accommodate both systems working based on age-group types and based on DOB. DOB being the standard. It is then to each implementor to specify which types he will support. DOB should not be mandatory though in passengerAbstract. we also can remove the list from the description 

yes, we could. This means small changes also in post /trip-offers-collection, post/offers (the idea was to have the embed values  related to offers and trips in the offer and trip search criterias, and only the ones directly related to tripoffeds in tripoffers. But I realise it may be less readable so ok to change