ERA request for reimbursement
reimbursement: refund or partial refund of a ticket that has not or only partially been used by the traveler. This could be the entire booking or anly a part of the passengers and a a part of the journey. The travellers have to provide evidence that they have not used the ticket (e.g. written confirmation from station or train staff). The reimbursement feature should only be used in case it can not be implemented via electronic data exchange between TCOs/Carriers on partially non-used tickets.